Update elf_to_header to support different options.
It is now possible to load the cluster code either as before or
loading the `.bin`. The `.bin` can either be loaded through a while
loop by `CVA6` or with the `udma_hyper`. The user always sees the
load_cluster_code function and selects the load type with different
defines.
 * No additional defines = standard load_code
 * CLUSTER_BIN embedds the code in the elf
 * CLUSTER_BIN + USE_HYPER loads the bin with the udma